More generic Authorization Code flow
Remove the hard dependency to the google cloud IAM and use the generic
OAuth2 Authorization Code Flow.
Configuration now happens completely through ENV variables.

Also removed the check if the user is still in the organization, because
it's not part of the oauth2 standard.
although it is a valid check and should be considered in the future.